Fix Windows test related to `shell` option (#185)
Fixes #56

Using shabangs on Windows currently only works in non-shell mode, but not in shell mode (when `shell` option is either `true` or a string). This is [done through `cross-spawn`](https://github.com/moxystudio/node-cross-spawn/blob/master/lib/parse.js#L32). I think the current behavior is the correct behavior.

The original test failed because it tried to fire a shabang with `cmd.exe`. Instead I've fired the binary through `node` which works.